+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased Changes
+------------------
+
+* Feature - Add support for inheritance. Aws Record models can now be extended using standard ruby inheritance (#80).
+
 2.8.0 (2022-10-12)
 ------------------